feat: 트라벨 슬라이더 좋아요 훅 사용 업데이트1
