+++ src/components/main/swiper/TravelSlider.tsx
@@ -1,23 +1,27 @@
 "use client";
 
+import { Likes } from "@/types/Likes.types";
 import { ApiInformation } from "@/types/Main";
+import { convertToHttps } from "@/utils/convertToHttps";
+import { useUserStore } from "@/zustand/userStore";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
-import { convertToHttps } from "@/utils/convertToHttps";
+import { useEffect, useState } from "react";
 import { useMediaQuery } from "react-responsive";
+import SwiperCore from "swiper";
 import { A11y, Autoplay, Grid } from "swiper/modules";
 import { Swiper, SwiperSlide } from "swiper/react";
+import { useHandleLikeButtonSwiper } from "./../../../hooks/detailpage/useHandleLikeButtonSwiper";
 import TravelSwiperSkeleton from "./TravelSwiperSkeleton ";
-import SwiperCore from "swiper";
-import { useCallback, useEffect, useState } from "react";
-import { Likes } from "@/types/Likes.types";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { useModal } from "@/context/modal.context";
-import { useUserStore } from "@/zustand/userStore";
-import { useLikes } from "@/hooks/detailpage/useLikes";
-import { createClient } from "@/utils/supabase/client";
 interface MainTravelSliderProps {
   travel: ApiInformation[];
+  contentId: string;
+  imageUrl: string;
+  contentTypeId: string;
+  title: string;
+  addr1: string;
+  tel: string;
+  userId: string;
 }
 interface ContextType {
   previousLikes: Likes[] | undefined;
@@ -27,219 +31,28 @@ export const MainTravelSlider: React.FC<MainTravelSliderProps> = ({ travel }) =>
   const isLgScreen = useMediaQuery({ minWidth: 1024 });
   const [swiper, setSwiper] = useState<SwiperCore | null>(null);
   const router = useRouter();
-  const supabase = createClient();
-  const queryClient = useQueryClient();
+
   const { id: userId } = useUserStore();
-  const modal = useModal();
+
   const item = travel[0] as ApiInformation;
-  const [likedStates, setLikedStates] = useState<Record<string, boolean>>({});
-  const { isPending, isError, data } = useLikes(item.contentid, userId);
 
-  const deleteMutation = useMutation({
-    mutationFn: async (userId: string) => {
-      const { error } = await supabase.from("Likes").delete().match({ user_id: userId, content_id: item.contentid });
-      if (error) {
-        throw new Error(error.message);
-      }
-    },
-    onSuccess: () => {
-      modal.open({
-        title: "알림",
-        content: (
-          <div className="text-center">
-            <p>좋아요가 취소되었습니다.</p>
-          </div>
-        ),
-      });
-      queryClient.invalidateQueries({ queryKey: ["likes", item.contentid] });
-    },
-    onError: (error) => {
-      modal.open({
-        title: "에러",
-        content: (
-          <div className="text-center">
-            <p>좋아요 취소 중 에러가 발생했습니다.</p>
-            <p>{error.message}</p>
-          </div>
-        ),
-      });
-    },
+  const { likedStates, isError, isPending, data, handleLikeButton } = useHandleLikeButtonSwiper({
+    travel,
+    contentId: item.contentid,
+    imageUrl: item.firstimage,
+    contentTypeId: item.contenttypeid,
+    title: item.title,
+    addr1: item.addr1,
+    tel: item.tel || "",
+    userId: userId || "",
   });
-
-  const addMutation = useMutation<Likes, Error, Partial<Likes>, ContextType>({
-    mutationFn: async (variables) => {
-      if (!userId) {
-        modal.open({
-          title: "알림",
-          content: (
-            <div className="text-center">
-              <p>로그인 후 좋아요를 누를 수 있습니다.</p>
-            </div>
-          ),
-        });
-
-        throw new Error();
-      }
-
-      const { data, error } = await supabase
-        .from("Likes")
-        .insert([
-          {
-            user_id: userId,
-            content_id: variables.content_id,
-            image_url: variables.image_url,
-            content_type_id: variables.content_type_id,
-            title: variables.title,
-            address: variables.address,
-            tel: variables.tel,
-          },
-        ])
-        .select()
-        .single();
-
-      if (error) {
-        throw new Error(error.message);
-      }
-
-      if (!data) {
-        modal.open({
-          title: "에러",
-          content: (
-            <div className="text-center">
-              <p>데이터가 없습니다.</p>
-            </div>
-          ),
-        });
-        throw new Error();
-      }
-
-      return data as Likes;
-    },
-
-    onMutate: async (variables) => {
-      await queryClient.cancelQueries({ queryKey: ["likes", variables.content_id] });
-      const previousLikes = queryClient.getQueryData<Likes[]>(["likes", variables.content_id]);
-
-      if (previousLikes) {
-        const newLikes = [...previousLikes, variables as Likes];
-        queryClient.setQueryData<Likes[]>(["likes", variables.content_id], newLikes);
-      } else {
-        queryClient.setQueryData<Likes[]>(["likes", variables.content_id], [variables as Likes]);
-      }
-
-      return { previousLikes };
-    },
-
-    onSettled: (variables) => {
-      modal.open({
-        title: "알림",
-        content: (
-          <div className="text-center">
-            <p>장소가 나의 공간에</p>
-            <p>추가되었습니다.</p>
-          </div>
-        ),
-      });
-      queryClient.invalidateQueries({ queryKey: ["likes", variables?.content_id] });
-    },
-
-    onError: (error, variables, context) => {
-      modal.open({
-        title: "에러",
-        content: (
-          <div className="text-center">
-            <p>좋아요 등록 중 에러가 발생했습니다.</p>
-            <p>{error.message}</p>
-          </div>
-        ),
-      });
-
-      if (context?.previousLikes) {
-        queryClient.setQueryData<Likes[]>(["likes", variables?.content_id], context.previousLikes);
-      }
-    },
-  });
-  const handleLikeButton = useCallback(
-    async (event: React.MouseEvent<HTMLButtonElement>) => {
-      event.stopPropagation();
-      const contentId = event.currentTarget.getAttribute("data-contentid"); // 클릭된 버튼의 contentId 가져오기
-
-      if (!contentId) {
-        modal.open({
-          title: "에러",
-          content: (
-            <div className="text-center">
-              <p>Content ID가 없습니다!</p>
-            </div>
-          ),
-        });
-
-        return;
-      }
-
-      try {
-        if (likedStates[contentId]) {
-          deleteMutation.mutate(userId!);
-        } else {
-          const item = travel.find((item) => item.contentid === contentId); // travel 배열에서 해당 contentId를 가진 item 찾기
-
-          if (!item) {
-            modal.open({
-              title: "에러",
-              content: (
-                <div className="text-center">
-                  <p>Item not found 에러</p>
-                </div>
-              ),
-            });
-            return;
-          }
-
-          addMutation.mutate({
-            user_id: userId!,
-            content_id: item.contentid,
-            image_url: item.firstimage,
-            content_type_id: item.contenttypeid,
-            title: item.title,
-            address: item.addr1,
-            tel: item.tel,
-          });
-        }
-
-        setLikedStates((prevStates) => ({
-          ...prevStates,
-          [contentId]: !prevStates[contentId],
-        }));
-      } catch (error) {
-        modal.open({
-          title: "에러",
-          content: (
-            <div className="text-center">
-              <p>좋아요 상태 업데이트를 실패했습니다.</p>
-              <p>{(error as Error).message}</p>
-            </div>
-          ),
-        });
-      }
-    },
-    [likedStates, addMutation, deleteMutation, userId, modal, travel],
-  );
 
   useEffect(() => {
     if (swiper) {
       swiper.update();
     }
   }, [isLgScreen, swiper]);
-  useEffect(() => {
-    if (data) {
-      const newLikedStates = travel.reduce((acc, item) => {
-        const result = data.find((like) => like.user_id === userId && like.content_id === item.contentid);
-        acc[item.contentid] = !!result;
-        return acc;
-      }, {} as Record<string, boolean>);
-      setLikedStates(newLikedStates);
-    }
-  }, [data, travel, userId]);
+
   if (isPending) {
     return <TravelSwiperSkeleton />;
   }
@@ -249,7 +62,7 @@ export const MainTravelSlider: React.FC<MainTravelSliderProps> = ({ travel }) =>
   }
 
   const likeImage =
-    data && data.find((like) => like.user_id === userId)
+    data && data.find((like: { user_id: string | null }) => like.user_id === userId)
       ? "/assets/images/successLikeIcon.svg"
       : "/assets/images/defaultLikeIcon.svg";
